requirements.txt
django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.3.3
python-dotenv==1.2.2
